2841: minimize header includes in View headers (#2874)
* 2841: clean up commands a bit

* 2841: pass raw command pointers to command state change notifications

* 2841: manage commands using unique_ptr to clear up ownership semantics

* 2841: fix bug in vertex tool

* 2841: clean up header includes

* 2841: add missing includes

* 2841: remove includes, use vector instead of set where possible

* 2841: fix missing includes

* 2841: remove unnecessary header includes

* 2841: remove unnecessary includes

* 2841: remove unnecessary includes

* 2841: remove unnecessary includes

* 2841: update vecmath

* 2841: add missing includes

* 2841: refactoring and documenting CommandProcessor, adding tests

* 2841: refactoring, adding more tests for CommandProcessor

* 2841: better support for nested transactions

* 2841: add more tests for command processor

* 2841: add vec_pop_back function

* 2841: use kdl::vec_pop_back method

* 2841: address review comments

* 2841: address review feedback

* 2841: fix missing include

* 2841: fix include

+++ common/src/View/Inspector.cpp
@@ -63,7 +63,7 @@ namespace TrenchBroom {
         }
 
         void Inspector::switchToPage(const InspectorPage page) {
-            m_tabBook->switchToPage(page);
+            m_tabBook->switchToPage(static_cast<int>(page));
         }
 
         bool Inspector::cancelMouseDrag() {
